Add keyboard layout for Urdu
The file `KeyboardTextsTable.java` states that it shouldn't be
edited manually. However, the tools required to generate that
file require the AOSP code and are not present in Simple Keyboard.
Therefore, I took the liberty of editing it manually. (The count of
texts at the end of the file (stated in comments) is also not
accurate.)